分离文件
